Add the concept of UpdateAll into the plugin loader
We had the gs_plugin_app_update() plugin vfunc to update a specific application,
and a gs_plugin_offline_update() vfunc to schedule a list of updates to be
installed offline. For us to support other packaging systems that supported
updating *all* the updatable GsApps live (for instance aptd) we need an update
method to take a list of GsApps, not a single object.

We can just rename the gs_plugin_offline_update() to gs_plugin_update() which
allows plugins that need to build a transaction of lots of packages rather than
one transaction for each GsApp. To avoid plugins having to each traverse the
list and check the management plugin, we also fall back to calling
gs_plugin_update_app() on plugins just like we do for refine() and refine_app().

This means that existing plugins that just schedule the install to be done just
have to check if any application is OFFLINE rather than OFFLINE_LIVE, and
plugins that don't issue a large transaction can continue to use the
gs_plugin_update_app() vfunc as it will only be run when the management plugin
is correct.

Many thanks to William Hua for prototyping all the different ways to do this,
and spending the time working out all the details.