Ensure valid AppStream files get the fake 'Addon' group applied
The Debian metadata extractor does not do this, and doing this "client-side"
is the correct way to assign DE-specific icons. This way we also always show
fonts with type=font and no categories, and we don't pollute appstream-glib
with a gnome-software-ism.

While we're breaking everything, remove the plural from the fake groups to
better match the desktop spec. Given appstream-glib and gnome-software disagree
on 'Fonts' and 'Font' and that both were written by the same person I think
it's about time to make this clearer.